Next, let’s run the query against our GraphQL API by calling the useQuery hook with our GET_POKEMON_INFO query. We use object destructuring to get the data, loading and error parts from the returned value of the useQuery hook.
If loading is true, this means data is still being fetched so we simply render the Loading… message code to indicate to users that data is still loading.

If an error occurs during the query, the error variable will have a value of true. In this case, we simply render an Error message.